+++ game/theater/start_generator.py
@@ -58,6 +58,7 @@ class ModSettings:
     a6a_intruder: bool = False
     a7e_corsair2: bool = False
     f4bc_phantom: bool = False
+    f9f_panther: bool = False
     f15d_baz: bool = False
     f_15_idf: bool = False
     f_16_idf: bool = False