+++ changelog.md
@@ -16,6 +16,7 @@
 * **[Modding]** Updated Community A-4E-C mod version support to 2.2.0 release.
 * **[Modding]** Added F/A-18E/F Super Hornet AI Tanker mod support (Chiller Juice Studios SuperBug Tanker AI version 1.4)
 * **[Modding]** Added VSN Super Étendard mod support (v2.5.5)
+* **[Modding]** Added F9F Panther mod support (version v2.8.7.101)
 * **[Modding]** Updated Irondome support to IDF Assets Pack V1.1, adding support for the David's Sling
 * **[Radios]** Added HF-FM band for AN/ARC-222
 * **[Radios]** Ability to define preset channels for radios on squadron level (for human pilots only)